move noun and verb filtering code into the connector
It's not exclusively a cli function, and we'll want to call it from
elsewhere, so it really belongs in the connector.
